test(controlplane): rename poll-with-backoff util to timers
The module now holds general timing helpers (sleep, computeDelay,
pollWithBackoff, retryWithBackoff), so timers.ts fits better than the
poll-specific name.
